llvm: Switch to trampoline-based jlcall annotations
As discussed extensively in #45057, when enabling LLVM's opaque pointer
support, we get significant miscompilations in jlcall callsites, because
calls with mismatching calling conventions are considered undefined behavior.
This implements Option D) from #45057, switching our jlcall callsites to use
a `julia.call` trampoline intrinsic instead. The lowering for this intrinsic
is essentially the same as the CC-based lowering before, except that the
callee is now of course the first argument rather than the actual callee.
Other than that, the changes are mostly mechanical.

Fixes #45057