Allow constprop to improve constant results to Bottom
This is consistent with what we would do for a non-constant return.
This helps the compiler dump less code onto LLVM in situations
where there is a big basic block that our constprop can see is
dead. That said, for that particular situation, it might be better
to have a constant propagation pass at the IR level that runs
after inlining. We can revisit that after #44660.