implement jl_rec_backtrace stackwalk for more platform configurations (#45110)
This adds support for most of the default configuration on most common
platforms (combinations of glibc/Darwin + AArch64/ARM/x86/x86_64). It
was already fully supported for Windows x86 and x86_64.

I would add support here for musl (since it is trivial), but they
explicitly prohibit via their headers, as a design choice, so I don't
know it is right to auto-detect support for it in the build.