Fix effect propagation for return_type
In addition to the TODO placeholder that was taininting all effects
for `return_type`, we were also accidentally picking up the effects
of the function we were analyzing. This mostly didn't show up as
an issue, because we were treating return_type as pure in optimization,
but it could prevent deletion of unused functions by incorrectly
tainting the effects.