Define Iterators.first as alias for Iterators.take
Two-argument Base.first is equivalent in meaning to a non-lazy Iterators.take. This commit
defines Iterators.first so that it mirrors Base.first in the same way that e.g. Base.reverse
is mirrored by Iterators.reverse. This should help with discoverability.

Technically this could be considered a breaking change for anyone who called "Base.first" as
"Iterators.first". However searching github the only results using that pattern are things
like people's personal collection of Advent of Code solutions, so breakage should be minimal
in reality.