Revert drive-by change-of-behavior in performance patch
The pull request #43816 is supposed to be a performance improvement
(only), but also changes behaviour: it's now possible for
`getdict!(dict, data)` to return with some elements of `data` not
a valid key in `dict`.

Maybe this is a desired improvement now that `data` also contains
metadata about taskid and threadid, but it's certainly breaking
at least one registered package (which I reported as #45361).

I propose we keep the performance improvement but revert the change
in behaviour. We can then discuss the behaviour's merit in a separate
pull request if desired, without the time pressure of the imminent
v1.8 release.