Make finalizer a built-in (#45423)
* Make finalizer a built-in

Split out from #45272. This is prepratory work towards adding
optimization passes that recognize this builtin. This PR adds
`Core.finalizer` with essentially the same interface as
`Base.finalizer`, but without the error checking or raw-C-pointer
feature. In future commits, the Core.finalizer interface will
likely expand slightly, but Base.finalizer will remain unchanged
and is the supported interface for this functionality.

* Update base/docs/basedocs.jl

Co-authored-by: Shuhei Kadowaki <40514306+aviatesk@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Ian Atol <ian.atol@juliacomputing.com>

Co-authored-by: Shuhei Kadowaki <40514306+aviatesk@users.noreply.github.com>
Co-authored-by: Ian Atol <ian.atol@juliacomputing.com>