inference: make `Conditional` and `InterConditional` slightly more symmetric
Especially the `.var::SlotNumber` field of `Conditional` is now renamed
to `.slot::Int` as like `InterConditional`. This change may also cut off
few allocations of `SlotNumber`.

We may want `Conditional` to propagate constraints about non-local binding
(like global references) in the future, but I think that won't happen that
soonish. For the meanwhile, it would be reasonable to keep the field
specific `SlotNumber`.