manage stackwalk/profile lock correctly
Previously we had the wrong number of them (it was not global) and it
was in the wrong shared library (it was part of codegen instead of
profiling).

The profile_round_robin_thread_order buffer was not allocated on the
thread that exclusively used it, so it sometimes could end up somewhat
awkwardly (in incorrectly) shared between 2 threads.