improve concrete-foldability of core math functions
By making more use of the `@assume_effects` annotation. This commit adds
several `:consistent` annotations to certain functions that access to
constant global table (e.g. `INV_2PI` and `J_TABLE`), because currently
our effect analysis  can't prove the consistency in the presence of
`@inbounds` and `:boundscheck`.

In the long term, we can revert this commit once we improve the effect
analysis so that it can properly reason about safe tuple indexing.