Refactor the meaning of NewSSAValue in IncrementalCompact (#45610)
Curently, uncompacted nodes in IncrementalCompact can refer
to old SSAValues (as regular SSAValues) and new-new SSAValues
(as NewSSAValues), but not post-compaction SSA values. This
is a bit of an odd missing middle and is useful when re-ordering
instructions from beyond the current insertion point. This PR
updates the definition of NewSSAValue to be able to refer
to both post-compaction SSAValues (using positive `id` values)
and new-new SSAValues (using negative `id` values), which also
makes their meaning independent of context (like OldSSAValue).

While we're at it, also clean up the `getindex` methods for
IncrementalCompact to return `Instruction` (as we did for IRCode).