BitVector: construct from Unsigned; l/r-shift(!), including in-place
Preface: there are two distinct concepts here -- a constructor for
`BitVector` from unsigned integers, and l/r-`shift[!]` methods which
match the corresponding shifts on bit indices.

When writing the code for `BitVector(x::Unsigned)`, I realized
that the shift operators (`>>`, `>>>`, `<<`) have the opposite
behavior w.r.t. the corresponding bitshifts. Odd, I thought, then
found [this](https://github.com/JuliaLang/julia/issues/30958) --
sufficient motivation to write the l/r-`shift[!]`
methods. Incidentally, these shifts happen to be quite useful when
testing the behavior of the construction-from-unsigned, which is why
these two concepts are initially being submitted as part of the same
PR.

Admittedly, I would normally have preferred to submit separate ideas
as separate PRs so as to facilitate rejecting one or the other; my
apologies for the inconvenience.

Miscellaneous: the implementation of the l/r-`shift[!]` methods
enables copy-free, in-place shifts on `BitVector`s, but it is not as
efficient as it possibly could be -- there are some partial passes
over the backing `Vector{UInt64}` which could probably be eliminated
by properly re-writing `copy_chunks_rshift!`. I also notice the
existence of a potentially dangerous but gainful opportunity for
`@inbounds` in `_msk_rtol!`.