`findextrema`: compute `findmin` and `findmax` in single pass
This is a simple extension of extant `findmin` and `findmax` methods.
Depending on context (cost of `f`; whether reduction is over dims;
size of array) the speedup increase is somewhere between 1.0-1.6 (no regressions).
Interestingly, I noticed but could not locate a `findextrema`; there
is some [mention](https://github.com/JuliaLang/julia/pull/7327) of it,
but nothing in Base. If it was deemed unworthy, please excuse this
errant PR.