uniform `Rational` numbers on the closed interval `[0,0]`
Upon receiving a method error for `rand(Rational{Int})`, I realized
that it would be possible to provide a sensible definition for a
uniform (discrete) distribution on the closed interval [0,1].  The
methods were designed with xoshiro RNGs (with which I have
familiarity) in mind. I need to look more closely at the
`MersenneTwister` implementation to determine whether a dispatch might
be warranted so as to get the "good bits"; I will get to this in the
next few days. In any case, I put this up
for comment.