Relax type of `PartialOpaque` source field
We type this as `::Method`, which is what's expected in Base. However,
Diffractor wants to be able to generate the source lazyily, so it needs
to be able to stash some information here to be able to perform
type inference before generating the method source. When that code
was originally written (pre-OpaqueClosure), this used a custom lattice
element, but `PartialOpaque` seems appropriate, since these do widen
to an OpaqueClosure and what's happening here is essentially just
that the `source` is being generated lazily.

We may want to revisit this if we switch the way that we represent
lattice elements, but for now, this seems like the path of least
resistance.