win: Propagate crashes to OS crash handlers
Similar in spirit to #46157. We'd like to collect crashdumps
when julia crashes on CI, but currently we just cleanly exit
the pocess in this case. By continuing exception handling into
the global crash handler, the OS's crash reporter gets invoked
and can dump out a minidump for us. In the future, we may want
to add our own crash reported, but this should hopefully help
debug crashes for the moment.