Allow elimination of `isbits` for non-const argument (#46330)
It should be sufficient for the _type_ of `x` to be known at compile
time to fully eliminate `isbits(x)`. However, the pevious code relied on
`isbits` being called on its argument during inference (by declaring it
`:total`) for eliminating it. But if the value of `x` was not known,
only its type, this would not work. The new implementation has the added
benefit of being simpler and DRYer (wrt. `isbitstype`).