More information on the parse function.
Providing more information on how the parse function works with floating point literals would be great, as I've ran into situations where I thought it intuitive to work with the Float32 numeric literal, but it throws an error. Providing a warning up-front about this behaviour and how to go around it would be great to inform users to take notice of such occurrences.