complex numbers from unsigned integers
From the docs, it's stated that using juxtaposition with identifiers as coefficients can provide convenient syntax for complex numbers. But then, using this syntax fails completely with all unsigned integers created from their numeric literals i.e. 0x, 0b, 0o. 

This is no trivial matter, and the docs completely made no word on it to inform people about this behaviour. So it's best if this is stated as a note to users, and also a way is provided on how to come around it.