More specificity about the isreal(x) function
isreal(x) seems to be too vague a function name when it's applied on complex numbers. My reasons are because:
1. We don't know if it's trying to ask if just the real part of the complex number is subtype of Real i.e. a real number.
2. We also don't know if its trying to ask if the whole number i.e. its real and imaginary part(if it has any) is a real number.

Number 2 reason is very important as we know technically speaking 1 is a real number and 2 + 3im is a complex number. So one would naturally expect isreal to return false for complex numbers (which in Julia does not). So you see it's kinda ambiguous.

I'm not proposing a function name change. What I'm saying is it will help if the isreal documentation provides more information on how it relates with complex numbers and probably more examples in regard to complex numbers, to at least lessen the ambiguities for users who care to check on the isreal definition before using it.

Julia> typeof(1 + 0im)
Complex{Int64}

Julia> isreal(1 + 0im)
true

So even though the imaginary part is 0, in Julia the number is still regarded as a complex number, which returns true as an argument to the isreal function, causing more confusion. So it will be best if it's stated explicitly, that this is the only way for complex numbers to return true from the isreal function.

To add to it, the docs examples and specificity on the relationship of isreal with complex numbers is too small - I believe the function is used more when working with complex numbers, so it makes sense to add a little info about how the isreal works with complex numbers and maybe one or two more examples where it will be helpful.