complex numbers should be treated equally in Julia
I keep coming across situations where complex numbers are thought of as a low use cases in Julia's docs. This is because they have far fewer examples, far fewer explanations on how stuffs are done with them, and far fewer explanations on how they interact with functions that can be used on them.

Julia is a general use programming language but it has a domain where it shines most and that's SCIENTIFIC COMPUTING. This is to say that inasmuch as anything is related to numbers, it should be treated equally in Julia (the Julia docs assumes we only use integers and floating points a lot).

If someone comes into Julia building an application or package that rigorously works with complex numbers and he is reading through the docs, he won't get the best experience as there are far fewer explanations on functions working with complex numbers and also far fewer examples of how the functions are used  on complex numbers.