compiler: Get rid of has_struct_const_info (#46564)
This was introduced by me in #39684, but the commit message has
no discussion as to why `has_struct_const_info` and
`has_nontrivial_const_info` are not the same function.

Moreover, I don't think the `isa(x, Conditional)` branch ever
did anything because even at the time of the commit, `tuple_tfunc`
applied `widenconditional` to all incoming argtypes. That leaves
`PartialTypeVar`, which seems reasonable to just include in the
definition of `has_nontrivial_const_info`, so that we may be
consistent between `tuple_tfunc` and `Expr(:new)`.