inlining: Add an option to not generate compilesig :invoke statements
In general, when we inline we may change which MethodInstance to invoke
to avoid over-specializing. This makes sense when generating IR to
be codegen-ed, but makes less sense when we are using inlining as
part of a code analysis pipeline, since the compilesig signature
will often have not been inferred yet. Add an optimizer option to
turn on/off generation of compilesig :invokes.