fix issue #46665, `prod(::Array{BigInt})` (#46667)
The way we were counting the number of bits was assigning a negative number
to `0`, which could lead to a negative total number of bits. Better to just
exit early in this case.
Also, the estimate was slightly off because we were counting the number of leading
zeros in the least significant limb, instead of the most significant.

(cherry picked from commit f7b4ebece6ec76289065c523bbc3a2856d3263ae)