Inlining/Finalizer-elimination cleanup refactor (#46700)
This started as an attempt to fix the broken test at [1], but I don't actually
think that test ever worked. I still intend to fix that, but this PR is prepratory,
merging code paths that basically did identical things, but were incomplete in some
cases. Also some general quality improvements. The test case in question is better
now (everything gets inlined and it's down to needing to prove that the finalizer
itself is inlineable and the mutable :new is eliminable), but not quite fixed.

[1] https://github.com/JuliaLang/julia/blob/master/test/compiler/inline.jl#L1306