Improve attribution of backedge-triggered invalidation
SnoopCompile attempts to attribute invalidations to specific causes,
but until now it has not generally been able to handle what it called
"delayed" invalidations, which arose when a MethodInstance backedge
wasn't valid anymore. This dumps more data to the reporting stream
and should allow SnoopCompile to assemble the full chain of causes.

This also adds invalidation of the backedges of methods that fail
to validate their external edges.