Use more of the actual inliner in finalizer inlining
Currently the special-case inliner for finalizer generated incorrect
code for unmatched sparams (which we had a test for, but which was
unfortunately broken) and varargs (which we did not have a test for).
Of course, we handle these cases correctly in actual inlining, so do
some light refactoring to share more code between the two inlining paths.