consistency on typemax and typemin docs
The `typemax` and `typemin` docs isn't consistent with each other on its examples and `see also` section (see reasons below).

While this isn't an issue when looking up the function with `help>` to work with integers, for floating-point types, its a big deal because:
- On the `typemax` examples, there is a very good example there with a great comment, pointing users on to look into `floatmax`, if `typemax` doesn't serves them; but the `typemin` doesn't have this - which isn't great, since both `typemax` and `typemin` have great use.
- The `see also` section on `typemax` point users to other available functions which might prove or introduce other points to consider if `typemax` isn't what they're looking for; for `typemin`, this isn't present.

I provided examples on integers, just so the function isn't thought of like working for only floating points.

Moreover, I feel like the current docs were written assuming users would always lookup `typemin` and `typemax` in `help>`, before using anyone of them. Most time this is true, but I think it's best not to assume.